Refactored into a version to be generated, invisible from outside
